+++ include/ann/layers/loss.h
@@ -9,9 +9,9 @@ namespace Layer {
 class Loss {
 public:
   // Calculate average loss from calculated output derived from member function
-  double mean();
+  float mean();
 
-  std::shared_ptr<const Math::Vector<double>> getOutput() const {
+  std::shared_ptr<const Math::Vector<float>> getOutput() const {
     return m_output;
   }
 
@@ -20,12 +20,12 @@ protected:
 
   // initialize member variables
   Loss(size_t batchNum)
-      : m_output{std::make_shared<Math::Vector<double>>(batchNum)} {};
+      : m_output{std::make_shared<Math::Vector<float>>(batchNum)} {};
 
   // Move constructor
   Loss(Loss &&other) : m_output{other.m_output} { other.m_output.reset(); }
 
-  std::shared_ptr<Math::Vector<double>> m_output{
-      std::make_shared<Math::Vector<double>>()};
+  std::shared_ptr<Math::Vector<float>> m_output{
+      std::make_shared<Math::Vector<float>>()};
 };
 } // namespace Layer
